+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Resolve textarea focus loss after interactive mode exit
 ---
 
 ## [0.37.0] - 2026-04-11